Consolidate main.c
