+++ src/Makefile.am
@@ -1,7 +1,7 @@
 AUTOMAKE_OPTIONS=subdir-objects
 AM_CFLAGS=-Wall -Wextra -pedantic-errors -Wno-unused-parameter -Werror
 bin_PROGRAMS=pick
-dist_pick_SOURCES=choice.c choice.h choices.c choices.h io.c io.h main.c \
-				  tty.c tty.h ui.c ui.h compat/queue.h compat/strlcpy.c \
-				  compat/strlcpy.h compat/strlcat.c compat/strlcat.h \
-				  compat/reallocarray.h compat/reallocarray.c
+dist_pick_SOURCES=main.c \
+		  compat/queue.h compat/strlcpy.c \
+		  compat/strlcpy.h compat/strlcat.c compat/strlcat.h \
+		  compat/reallocarray.h compat/reallocarray.c
