--- src/ui.h
@@ -1,9 +1,0 @@
-#ifndef UI_H
-#define UI_H
-
-#include "choice.h"
-#include "choices.h"
-
-struct choice	*ui_selected_choice(struct choices *, char *, int);
-
-#endif /* UI_H */
